+++ esp-hal/src/analog/adc/mod.rs
@@ -97,7 +97,6 @@ pub enum AdcCalSource {
     /// Use Ground as the calibration source
     Gnd,
     /// Use Vref as the calibration source
-    #[cfg(not(esp32h2))]
     Ref,
 }
 
@@ -227,7 +226,7 @@ impl<ADCI> AdcCalScheme<ADCI> for () {
 }
 
 /// A helper trait to get access to ADC calibration efuses.
-#[cfg(not(any(esp32, esp32s2, esp32h2)))]
+#[cfg(not(any(esp32, esp32s2)))]
 trait AdcCalEfuse {
     /// Get ADC calibration init code
     ///