+++ esp-hal/src/analog/adc/riscv.rs
@@ -8,10 +8,9 @@ cfg_if::cfg_if! {
     }
 }
 
-#[cfg(not(esp32h2))]
 pub use self::calibration::*;
 use super::{AdcCalSource, AdcConfig, Attenuation};
-#[cfg(any(esp32c2, esp32c3, esp32c6))]
+#[cfg(any(esp32c2, esp32c3, esp32c6, esp32h2))]
 use crate::efuse::Efuse;
 use crate::{
     Async,
@@ -192,6 +191,12 @@ impl super::CalibrationAccess for crate::peripherals::ADC1<'_> {
             AdcCalSource::Gnd => regi2c::ADC_SAR1_ENCAL_GND.write_field(enable as _),
             #[cfg(not(esp32h2))]
             AdcCalSource::Ref => regi2c::ADC_SAR1_ENCAL_REF.write_field(enable as _),
+            // For the ESP32-H2 ground and internal reference voltage are mutually exclusive and
+            // you can toggle between them.
+            //
+            // See: <https://github.com/espressif/esp-idf/blob/5c51472e82a58098dda8d40a1c4f250c374fc900/components/hal/esp32h2/include/hal/adc_ll.h#L645>
+            #[cfg(esp32h2)]
+            AdcCalSource::Ref => regi2c::ADC_SAR1_ENCAL_GND.write_field(!enable as _),
         }
     }
 }
@@ -404,7 +409,7 @@ impl<ADCI> InterruptConfigurable for Adc<'_, ADCI, Blocking> {
     }
 }
 
-#[cfg(all(adc1, not(esp32h2)))]
+#[cfg(adc1)]
 impl super::AdcCalEfuse for crate::peripherals::ADC1<'_> {
     fn init_code(atten: Attenuation) -> Option<u16> {
         Efuse::rtc_calib_init_code(1, atten)